+++ libMx/d_vp9000.c
@@ -485,7 +485,7 @@ mxd_vp9000_motor_is_busy( MX_MOTOR *motor )
 	MX_VP9000_MOTOR *vp9000_motor;
 	char c, c2;
 	int i;
-	unsigned long num_input_bytes_available;
+	uint32_t num_input_bytes_available;
 	mx_status_type mx_status;
 
 	mx_status = mxd_vp9000_get_pointers(motor,
